app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
google-genai>=1.0.0
pydantic-settings==2.2.1
python-dotenv==1.0.1